requirements.txt
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
openai==1.24.0
uvicorn
PyPi
fastapi